+++ lib/QECCore/test/test_codes.jl
@@ -82,4 +82,8 @@
     @testset "Quantum Tillich-Zemor Codes" begin
         include("codes/tillichzemor.jl")
     end
+    
+    @testset "Circulant Bivariate Bicycle Codes" begin
+        include("codes/generalized_circulant_bivariate_bicycle.jl")
+    end
 end